+++ test/token_test.rb
@@ -7,24 +7,12 @@ class TokenTest < Minitest::Test
     assert_equal expected, Mailkick.secret_token.unpack1("h*")
   end
 
-  def test_message_verifier_v1
-    message = "BAhbCTBpBkkiCVVzZXIGOgZFRkkiCnNhbGVzBjsAVA==--93b04b720fd4103e3826a3e7f02ff0c4b6b63a44"
-    assert_equal [nil, 1, "User", "sales"], Mailkick.message_verifier.verify(message)
-  end
-
-  def test_message_verifier_v2
+  def test_message_verifier
     message = "W251bGwsMSwiVXNlciIsInNhbGVzIl0=--68e6af4bc88e9910a912da36f779c349c4ac661d"
     assert_equal [nil, 1, "User", "sales"], Mailkick.message_verifier.verify(message)
   end
 
-  def test_custom_token_v1
-    with_secret_token("1" * 128) do
-      message = "BAhbCTBpBkkiCVVzZXIGOgZFRkkiCnNhbGVzBjsAVA==--e37be0a9c71177dd4d2dcb078e8bf744bab5b955"
-      assert_equal [nil, 1, "User", "sales"], Mailkick.message_verifier.verify(message)
-    end
-  end
-
-  def test_custom_token_v2
+  def test_custom_token
     with_secret_token("1" * 128) do
       message = "W251bGwsMSwiVXNlciIsInNhbGVzIl0=--fb88c71ff1d08d86ffd05b19674c162588aad283"
       assert_equal [nil, 1, "User", "sales"], Mailkick.message_verifier.verify(message)