+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Dropped support for legacy opt-outs
 - Dropped support for legacy secret token generation and serialization
+- Dropped support for legacy SendGrid service
 
 ## 1.4.0 (2025-04-03)