+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 2.0.0 (unreleased)
 
 - Added `prefix` option to `has_subscriptions`
-- Switched to suppressions API for Postmark
+- Switched to suppressions API for Postmark and changed default stream to `broadcast`
 - Dropped support for legacy opt-outs
 - Dropped support for legacy secret token generation and serialization
 - Dropped support for legacy SendGrid service